chore: Update optional dependencies.
