+++ pyproject.toml
@@ -28,9 +28,13 @@ importlib-metadata = {version = "^4.0", python = "<3.9"}
 ansys-fluent-core = "~=0.34.0"
 pyvista = ">=0.44.0"
 matplotlib = ">=3.6.0"
-pyvistaqt = "~=0.11.1"
-pyside6 = "~=6.8.1"
-imageio= ">=2.36.1"
+pyvistaqt = { version = "~=0.11.1", optional = true }
+pyside6 = { version = "~=6.8.1", optional = true }
+imageio = { version = ">=2.36.1", optional = true }
+
+[tool.poetry.extras]
+interactive = ["pyvistaqt"]
+single-window = ["pyside6"]
 
 [tool.poetry.urls]
 "Documentation" = "https://visualization.fluent.docs.pyansys.com/"
