+++ src/ansys/fluent/visualization/graphics/single_qt_windows.py
@@ -1,0 +1,70 @@
+# Copyright (C) 2022 - 2025 ANSYS, Inc. and/or its affiliates.
+# SPDX-License-Identifier: MIT
+#
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+"""Module defining qt single window to add graphics."""
+
+from PySide6.QtWidgets import (
+    QApplication,
+    QMainWindow,
+    QTabWidget,
+    QVBoxLayout,
+    QWidget,
+)
+
+
+# Define the main window
+class MainWindow(QMainWindow):
+    def __init__(self):
+        super().__init__()
+        self.setWindowTitle("PyFluent Visualization Plots")
+        screen = QApplication.primaryScreen().availableGeometry()
+        width = int(screen.width() * 0.75)
+        height = int(screen.height() * 0.75)
+        self.setGeometry(
+            (screen.width() - width) // 2,  # Center X
+            (screen.height() - height) // 2,  # Center Y
+            width,
+            height,
+        )
+        self.tabs = QTabWidget()
+        self.setCentralWidget(self.tabs)
+        self.tabs.setMovable(True)
+        self.plotters = []
+
+    def _add_tab(self, plotter, title="-"):
+        tab = QWidget()
+        layout = QVBoxLayout()
+
+        self.plotters.append(plotter)
+        if plotter:
+            layout.addWidget(plotter.interactor)
+        tab.setLayout(layout)
+
+        # Add tabs with PyVista BackgroundPlotters
+        self.tabs.addTab(tab, f"PyViz ({title})")
+
+    def closeEvent(self, event):
+        """Ensure proper cleanup of plotter instances on window close."""
+        for plotter in self.plotters:
+            plotter.close()  # Properly close each PyVista plotter
+        event.accept()
+        global _qt_window
+        _qt_window = None
